Chip OS v1.0 + Stratmict Kernel v1.0
GUESS WHAT?? ITS HERE!!! Chip OS v1.0 is the first official verison of Chip OS. It still doesn't have keyboard support yet, but there's so much that it should be considered a complete OS. It runs programs from a "prg" folder, instead of being built into the command prompt. All you have to do to add a command is to add a .py or .mpy file to the "prg" folder.